\name{SimBindProfiles-package}
\alias{SimBindProfiles-package}
\alias{SimBindProfiles}
\docType{package}
\title{
  Similar Binding Profiles, identifies common and unique regions in array genome tiling array data
}
\description{
  SimBindProfiles, identifies common and unique binding regions in genome tiling array data. 
  This package does not rely on peak calling, but directly compares binding profiles 
  processed on the same array platform. It implements a simple threshold approach, 
  thus allowing retrieving commonly and differentially bound regions between 
  datasets as well as events of compensation and increased binding. 
}

\details{
\tabular{ll}{
  Package: \tab SimBindProfiles\cr
  Type: \tab Package\cr
  License: \tab GPL-3\cr
}
}

\author{
  Bettina Fischer \email{bef22@cam.ac.uk}
  Enrico Ferrero \email{ef300@cam.ac.uk}
  Robert Stojnic \email{rs550@cam.ac.uk}
  Steve Russell \email{sr120@cam.ac.uk}
  Maintainer: Bettina Fischer \email{bef22@cam.ac.uk}
}

\keyword{ package }
